# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "vLLM CI Dashboard API",
    "version": "1.0.0",
    "description": "Public read-only API backing https://ci.vllm.ai. All documented GET endpoints are unauthenticated and return JSON; errors use `{ \"error\": string }`. Responses are cached via `Vercel-CDN-Cache-Control`; treat data as eventually consistent (30s-10min staleness)."
  },
  "servers": [
    {
      "url": "https://ci.vllm.ai",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/eval/samples": {
      "get": {
        "operationId": "get-eval-samples",
        "summary": "Per-sample eval results",
        "description": "",
        "tags": [
          "Eval"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "build_id",
            "schema": {
              "type": "string",
              "description": "Required; 400 when missing"
            },
            "required": false,
            "description": "Required; 400 when missing"
          },
          {
            "in": "query",
            "name": "task",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "workload",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "correct",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Omit to return all samples"
            },
            "required": false,
            "description": "Omit to return all samples"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number",
              "description": "Clamped to 1-5000 (default 200)"
            },
            "required": false,
            "description": "Clamped to 1-5000 (default 200)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/github/commit": {
      "get": {
        "operationId": "get-github-commit",
        "summary": "Look up a commit in vllm-project/vllm",
        "description": "Proxies the GitHub API for a commit in vllm-project/vllm.",
        "tags": [
          "GitHub"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sha",
            "schema": {
              "type": "string",
              "description": "Commit SHA matching ^[0-9a-f]{7,40}$ (case-insensitive); 400 if missing/invalid, 404 if not found"
            },
            "required": false,
            "description": "Commit SHA matching ^[0-9a-f]{7,40}$ (case-insensitive); 400 if missing/invalid, 404 if not found"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/gpu": {
      "get": {
        "operationId": "get-gpu",
        "summary": "Get GPU metrics (legacy combined endpoint)",
        "description": "History plus latest per-GPU readings in one response; kept for backward compatibility, superseded by `/api/gpu/history` and `/api/gpu/latest`.",
        "tags": [
          "GPU"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "schema": {
              "type": "number",
              "description": "Clamped to 1-720, non-numeric falls back to 24 (default 24)"
            },
            "required": false,
            "description": "Clamped to 1-720, non-numeric falls back to 24 (default 24)"
          },
          {
            "in": "query",
            "name": "hostname",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/gpu/history": {
      "get": {
        "operationId": "get-gpu-history",
        "summary": "Get GPU utilization history",
        "description": "Time-bucketed aggregates read from raw snapshots for windows up to 6 hours, from the 5-minute rollup for longer windows.",
        "tags": [
          "GPU"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "schema": {
              "type": "number",
              "description": "Clamped to 1-720, non-numeric falls back to 24 (default 24)"
            },
            "required": false,
            "description": "Clamped to 1-720, non-numeric falls back to 24 (default 24)"
          },
          {
            "in": "query",
            "name": "hostname",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "operationId": "get-jobs",
        "summary": "Job failure ranking and duration stats",
        "description": "Without startDate/endDate, only jobs with runs in the last 7 days are included.",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pipeline",
            "schema": {
              "type": "string",
              "description": "default \"CI\""
            },
            "required": false,
            "description": "default \"CI\""
          },
          {
            "in": "query",
            "name": "branch",
            "schema": {
              "type": "string",
              "description": "default \"main\""
            },
            "required": false,
            "description": "default \"main\""
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
...
}